feat(django 1.9): Move jinja2 tags out of templatetags folders
Django 1.9 now automatically imports all models in templatetags. This causes ImportErrors to be raised for users who do not have jinja2 installed.
