+++ docs/advanced_topics/jinja2.rst
@@ -9,6 +9,10 @@ Wagtail supports Jinja2 templating for all front end features. More information 
 Configuring Django
 ==================
 
+.. versionchanged:: 1.3
+
+    Jinja2 tags were moved from "templatetags" into "jinja2tags" to separate them from Django template tags.
+
 Django needs to be configured to support Jinja2 templates. As the Wagtail admin is written using regular Django templates, Django has to be configured to use both templating engines. Wagtail supports the Jinja2 backend that ships with Django 1.8 and above. Add the following configuration to the ``TEMPLATES`` setting for your app:
 
 .. code-block:: python
@@ -20,9 +24,9 @@ Django needs to be configured to support Jinja2 templates. As the Wagtail admin 
             'APP_DIRS': True,
             'OPTIONS': {
                 'extensions': [
-                    'wagtail.wagtailcore.templatetags.jinja2.core',
-                    'wagtail.wagtailadmin.templatetags.jinja2.userbar',
-                    'wagtail.wagtailimages.templatetags.jinja2.images',
+                    'wagtail.wagtailcore.jinja2tags.core',
+                    'wagtail.wagtailadmin.jinja2tags.userbar',
+                    'wagtail.wagtailimages.jinja2tags.images',
                 ],
             },
         }
