+++ wagtail/wagtailadmin/jinja2tags.py
@@ -3,7 +3,7 @@ from __future__ import absolute_import
 import jinja2
 from jinja2.ext import Extension
 
-from .wagtailuserbar import wagtailuserbar
+from .templatetags.wagtailuserbar import wagtailuserbar
 
 
 class WagtailUserbarExtension(Extension):
