+++ wagtail/wagtailimages/jinja2tags.py
